vectorizer/requirements.txt
fastapi
uvicorn[standard]
chromadb
sentence-transformers
numpy
pydantic
python-dotenv
